+++ blurt_cpp_80211/crc.h
@@ -4,14 +4,14 @@
 
 class CRC {
 private:
-    int L, s, M;
+    size_t L, s, M;
     uint32_t m;
     std::vector<uint32_t> lut;
     bitvector b;
     bitvector correct_remainder;
     void remainder_slow(const bitvector &a, bitvector &output);
     uint32_t remainder_fast(const bitvector &a, bitvector &output) const;
-    void lut_bootstrap(const bitvector &new_b, int new_L);
+    void lut_bootstrap(const bitvector &new_b, size_t new_L);
 public:
     CRC();
     void FCS(const bitvector &calculationFields, bitvector &output) const;
