+++ blurt_cpp_80211/interleave.h
@@ -1,18 +1,18 @@
 #include "blurt.h"
 
-void interleave_permutation(int Ncbps, int Nbpsc, std::vector<int> &j);
-void interleave_inverse_permutation(int Ncbps, int Nbpsc, std::vector<int> &k);
+void interleave_permutation(size_t Ncbps, size_t Nbpsc, std::vector<size_t> &j);
+void interleave_inverse_permutation(size_t Ncbps, size_t Nbpsc, std::vector<size_t> &k);
 
 template <class T>
-void interleave(const std::vector<T> &input, int Ncbps, int Nbpsc, bool reverse, std::vector<T> &output) {
-    std::vector<int> p;
+void interleave(const std::vector<T> &input, size_t Ncbps, size_t Nbpsc, bool reverse, std::vector<T> &output) {
+    std::vector<size_t> p;
     if (reverse)
         interleave_permutation(Ncbps, Nbpsc, p);
     else
         interleave_inverse_permutation(Ncbps, Nbpsc, p);
-    int N = input.size();
+    size_t N = input.size();
     output.resize(input.size());
-    for (int i=0; i<N; i+=Ncbps)
-        for (int j=0; j<Ncbps; j++)
+    for (size_t i=0; i<N; i+=Ncbps)
+        for (size_t j=0; j<Ncbps; j++)
             output[i+j] = input[i+p[j]];
 }
