+++ blurt_cpp_80211/scrambler.h
@@ -4,10 +4,10 @@
 
 class Scrambler {
 private:
-    int state;
+    uint32_t state;
 public:
-    Scrambler(int state);
+    Scrambler(uint32_t state);
     int next();
-    static void scramble(const bitvector &input, int multipleOf, bitvector &output, int scramblerState=0x7f);
+    static void scramble(const bitvector &input, size_t multipleOf, bitvector &output, uint32_t scramblerState=0x7f);
 };
 #endif
